Finished?
Seperated button creation to menu modules, allowing for the customization of button placement and menu placement. Added styles, which allow you to make buttons look consistent through the whole card. Add a refresh styles context menu, updates the menu modules on the table to the new style. Removed the propertyToggleBuilder factory from the encoder. Modules must now call enc.call("APItoggleProperty",{obj,propID}) on themselves.
It is recommend to call this inside the activateFunc target param.
Removed CallOnActivate Property param. ActivateFunc param is now required.
Properties now have a tags="list,of,tags" param, this is used by the Menus to decide what buttons belong in themselves. Similarly, the visible param is a boolean for should this property be even drawn in the first place.
Players can now be encoded, allowing for values to be tied to players.
Tools No longer Exist as a seperate value, they are now tracked as a property.